+++ packages/bp/src/core/messaging/messaging-router.ts
@@ -2,8 +2,8 @@ import * as sdk from 'botpress/sdk'
 import { HTTPServer } from 'core/app/server'
 import { CustomRouter } from 'core/routers/customRouter'
 import { Router } from 'express'
-import { MessagingLegacy } from './legacy'
 import { MessagingService } from './messaging-service'
+import { MessagingLegacy } from './subservices/legacy'
 
 export class MessagingRouter extends CustomRouter {
   private legacy: MessagingLegacy
